// src/client/shared/services/field-validator/field-validator.enum.ts
export enum FieldValidatorEnum {
  WEAK_PASSWORD = 'VALIDATE.WEAK_PASSWORD',
  INVALID_EMAIL_ADDRESS = 'VALIDATE.INVALID_EMAIL_ADDRESS',
  GENERAL_ERROR = 'VALIDATE.GENERAL_ERROR',
  INVALID = 'VALIDATE.INVALID',
  INVALID_URL = 'VALIDATE.INVALID_URL',
  INVALID_ID = 'VALIDATE.INVALID_ID',
  INVALID_NAME = 'VALIDATE.INVALID_NAME',
  INVALID_ADDRESS = 'VALIDATE.INVALID_ADDRESS',
  INVALID_COORDINATE = 'VALIDATE.INVALID_COORDINATE',
  INVALID_TIMESTAMP = 'VALIDATE.INVALID_TIMESTAMP',
  INVALID_PHONE_NUMBER = 'VALIDATE.INVALID_PHONE_NUMBER',
  INVALID_VERIFICATION_CODE = 'VALIDATE.INVALID_VERIFICATION_CODE',
  INVALID_TYPE = 'VALIDATE.INVALID_TYPE',
  MAX = 'VALIDATE.MAX',
  MIN = 'VALIDATE.MIN',
  MIN_DATE = 'VALIDATE.MIN_DATE',
  MAX_DATE = 'VALIDATE.MAX_DATE',
  DATE_BETWEEN = 'VALIDATE.DATE_BETWEEN',
  LENGTH = 'VALIDATE.LENGTH',
  MAX_LENGTH = 'VALIDATE.MAX_LENGTH',
  MIN_LENGTH = 'VALIDATE.MIN_LENGTH',
  MISMATCH = 'VALIDATE.MISMATCH',
  REQUIRED = 'VALIDATE.REQUIRED',
  MISMATCH_PASSWORDS = 'VALIDATE.MISMATCH_PASSWORDS',
  MISSING_UPPERCASE_CHARACTER = 'VALIDATE.MISSING_UPPERCASE_CHARACTER',
  MISSING_LOWERCASE_CHARACTER = 'VALIDATE.MISSING_LOWERCASE_CHARACTER',
  MISSING_NUMERIC_CHARACTER = 'VALIDATE.MISSING_NUMERIC_CHARACTER',
  MISSING_SPECIAL_CHARACTER = 'VALIDATE.MISSING_SPECIAL_CHARACTER',
  ALREADY_EXISTS = 'VALIDATE.ALREADY_EXISTS',
}